move routers into subdir
