+++ routers/connect.js
@@ -1,5 +1,5 @@
 var connect = require('connect');
-var harness = require('./harness');
+var harness = require('../harness');
 
 var routes = connect.router(function(app) {
   harness.addRoutes(function(name, handler) {
