+++ routers/http-hash.js
@@ -1,6 +1,6 @@
 var router = require('http-hash')();
 var http = require('http');
-var harness = require('./harness');
+var harness = require('../harness');
 
 
 harness.addRoutes(function(name, handler) {
