+++ routers/light-router.js
@@ -1,6 +1,6 @@
 var http = require('http');
 var router = require('light-router');
-var harness = require('./harness');
+var harness = require('../harness');
 
 harness.addRoutes(function(name, handler) {
   router.get('/' + name + '/:id', handler);
