+++ routers/rawhttp.js
@@ -1,4 +1,4 @@
 var http = require('http');
-var harness = require('./harness');
+var harness = require('../harness');
 
 var server = http.createServer(harness.makeHandler('default')).listen(harness.port);
